aarch64: Fix gcc.target/aarch64/sve/loop_add_6.c for LLP64 targets
Even though the aarch64-mingw32 support has not been committed yet,
we should fix some of the testcases. In this case gcc.target/aarch64/sve/loop_add_6.c
is easy to fix. We should use __SIZETYPE__ instead of `unsigned long` for the variables
that will be used for pointer plus.

Committed as obvious after a quick test on aarch64-linux-gnu.

gcc/testsuite/ChangeLog:

	PR testsuite/114177
	* gcc.target/aarch64/sve/loop_add_6.c: Use __SIZETYPE__ instead
	of `unsigned long` for index and offset variables.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>