[PATCH v4 4/4] Output S_COMPILE3 symbol in CodeView debug section
Outputs the S_COMPILE3 symbol in the CodeView .debug$S debug section.
The DEBUG_S_SYMBOLS block added here makes up pretty much everything
that isn't data structures or line numbers; we add the S_COMPILE3 symbol
here to start it off.

This is a descriptive bit, the most interesting part of which is the
version of the compiler used.

gcc/
	* dwarf2codeview.cc (DEBUG_S_SYMBOLS): Define.
	(S_COMPILE3, CV_CFL_80386, CV_CFL_X64): Likewise.
	(CV_CFL_C, CV_CFL_CXX): Likewise.
	(SYMBOL_START_LABEL, SYMBOL_END_LABEL): Likewise.
	(start_processor, language_constant): New functions.
	(write_compile3_symbol, write_codeview_symbols): Likewise.
	(codeview_debug_finish): Call write_codeview_symbols.