c++/modules: Ensure all partial specialisations are tracked [PR114947]
Constrained partial specialisations aren't all necessarily tracked on
the instantiation table.  The modules code uses a separate
'partial_specializations' table to track them instead to ensure that
they get walked and emitted when emitting a module, but currently this
does not always happen.

The attached testcase fails in two ways.  First, because the partial
specialisation is just a declaration (and not a definition),
'set_defining_module' never ends up getting called on it and so it never
gets added to the partial specialisation table.  We fix this by ensuring
that when partial specializations are created they always get added, and
so we never miss one. To prevent adding partial specialisations multiple
times we split this out as a new function.

The second way it fails is that when exporting the primary interface for
a module with partitions, we also re-walk the specializations of all
imported partitions to merge them into a single BMI.  So this patch
ensures that after calling 'match_mergeable_specialization' we also
ensure that if the name came from a partition it gets added to the
specialization table so that a dependency is correctly created for it.

	PR c++/114947

gcc/cp/ChangeLog:

	* cp-tree.h (set_defining_module_for_partial_spec): Declare.
	* module.cc (trees_in::decl_value): Track partial specs coming
	from partitions.
	(set_defining_module): Don't track partial specialisations here
	anymore.
	(set_defining_module_for_partial_spec): New function.
	* pt.cc (process_partial_specialization): Call it.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/partial-4_a.C: New test.
	* g++.dg/modules/partial-4_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>